requirements.txt
Flask==2.2.2
slack_sdk
firebase_admin
python-dotenv